renamed signin and signup pages. added not-found page. added not-found page to routes
