+++ src/pages/index.js
@@ -2,8 +2,9 @@
 
 import Home from './home';
 import Video from './video';
-import SignIn from './signin';
-import SignUp from './signup';
+import SignIn from './sign-in';
+import SignUp from './sign-up';
+import NotFound from './not-found';
 
 
 
@@ -11,5 +12,6 @@ export {
   Home,
   Video,
   SignIn,
-  SignUp
+  SignUp,
+  NotFound
 };
