+++ src/pages/not-found/index.jsx
@@ -1,0 +1,9 @@
+const NotFound = () => {
+  return (
+    <div>NotFound</div>
+  );
+};
+
+
+
+export default NotFound;
